Two related refactorings and bugfixes:
1. Cleaned up the Swing session context interface and implementations so they have a more useful set of session creation methods, which optionally show the architect frame.

as a result of this API change, the load file worker is back in action (in fact, it's been transplanted into a public static method of the OpenProjectAction).

2. Added an initGUI method to the swing session interface, and made the old init() method of the swing session (formerly of the ArchitectFrame itself) public.  Now it is completely optional to call, and if you do call it, it creates and configures the Architect Frame properly.

A few unit tests needed minor tweaks to deal with the new session and session context APIs.  They haven't been changed materially.