Made some modifications to account for the fact that Windows uses backslashes while Unix style pathnames use forward slashes. In particular, when composing the absolute build pathname from a relative pathname.
Also changed the way a Windows pathname with a drive letter is detected. It will only consider the single drive letter followed by a ':\' as an absolute path, as having a string of letters followed by ':\' doesn't seem to be considered an absolute path when creating the directories under Windows.