Fixed NPE for the following case:
If you pass in a list of tables and one or more of those tables has a relationship to another table not in the original list you passed in.  The simplest case to think of is when you pass in a list with only one table in it, and that table has an exported key.

The fix is to simply ignore tables not in the original list, so that the finishOrder will always contain exactly the same set of tables originally given.  This will lead to incorrect results if one of the omitted tables links back up with a table in the set, but then that's the calling code's fault. :)