This should fix bug #1431.
Basically, the problem was SQLRelationship.RelationshipManager.ensureInMapping was making changes that would only work on the assumption that in a relationship mapping, the primary key table and foreign key tables were not the same, which would not be the case in a self-referencing table.

This resulted in the method performing some operations that would not make sense in a self-referencing relationship, such as adding the foreign key column (since it's already there), and if the relationship is non-identifying, making sure the foreign key column is not in the primary key, which it would do if in a self-referencing relationship, we moved the foreign key into the primary key. (Note that this is part of another bug, where if we moved the foreign key into the primary key, the relationship is still marked as 'non-identifying).