To discover a profile function descriptor, the profiling system now requires a non-null parent for ColumnProfileResult. So now we provide one!
We also realized that when the PFD is missing for a certain data type (for example, varchar), then the discovery process is performed for EVERY COLUMN of that type! We put in a simple optimization to get around this for a single instance of RemoteDatabaseProfileCreator, but a further optimization would be to save the discovered PFD back to the PL.INI database type entry.

This change fixes the problem described in forum post 1871.