Fixed bug reported in forum post 1899 (cascading new primary key columns through child tables to grandchild tables and beyond was not working).
The problem: magic should not have been disabled while ensureInMapping() was doing its thing, but the fact that it was (incorrectly) disabled was hiding the fact that it wasn't properly placing the new FK column in the primary key of the child (for identifying relationships).

So, magic remains enabled for this operation now, and the fk col is properly inserted (or promoted) to the primary key.