Fixed bug 1477. Implements synchronization between selection on the dbTree and the playpen.
Removed select and show for now, which will be re-implemented next.
Deletion on dbTree throws an exception occasionally but will be gone when bug 1193 is fixed.

There's minor "bug" where ctrl select does not deselect a relationship on the dbTree. Selecting other objects deselects it as appropriate.
TODO: reimplement select and show in pp.