Big overhaul to the delete action: It's now based on the tree selection, not the playpen selection. This change was necessary in order to allow deleting indexes, which are not represented on the playpen, so they can't be selected there.
I also greatly simplified the code that figures out what to delete and not delete, and factored it out into a private method so it doesn't need to be duplicated in setupAction() anymore.

This change allowed me to remove an exception handler which evidently could not have been triggered before. I removed the corresponding messages from the localized files.

Finally, I changed the business model for SQLTable so the locked column check happens in the columns folder instead of the table. Before now, you could have deleted a locked column simply by deleting it from the folder instead of the table's deleteColumn() convenience method (which is how the delete action does it now). This is fixed, but it meant I had to make LockedColumnException unchecked.

In short, there is now way less code which does way more. :)