Relocated the ETL Lineage watcher from inside the PlayPen to a top-level class in the core package, because it enforces a business model invariant--it's not just a feature of the GUI.
This change also paves the way for allowing multiple playpens to exist in the same session, which is something we're already doing with the OLAP cube designer.

As a side effect, we had to relocate some messages into the core package, so there's now a Messages class in there, but it's only used from the ETL enforcer at the moment.

As a further side effect, we had to make the user prompter behaviour in the core session a little more flexible (so the Swing session can handle any prompting that's attempted directly on the core session it wraps)