Removed responsibility for building the popup menu from the PlayPen itself, and created an interface that the playpen uses to ask for a popup menu.
This new interface is implemented in a new helper class, RelationalPlayPenFactory, which is the new class responsible for setting up the increasingly generic PlayPen so it works just like the old relational playpen did.

The whole point of this is that we want to be able to use the PlayPen for OLAP modeling without coupling it to both SQLObjects and OLAPObjects.