Changed the way the session keeps OLAP schemas: Now there is an OLAPRootObject, which contains multiple OLAPSessions, each having one schema.
The need for the RootObject is that we need the ability to ask for event notification when new schemas are added and existing ones are removed.

The need for the OLAPSession objects between the root and the schema is that we need extra information that's not possible to store in the schema. Currently, this includes which SQLDatabase the schema uses.

There is also a fix for the tree model: It's now capable of specifying change events on the Schema object (in Swing, change events on the tree's root object are a special case)

Finally, we're committing a SchemaEditPanel in this commit because we were working on it at the time the need for all these other things became apparent.