Further drag'n'drop fun: Now all OLAPPanes support drop of the appropriate item type, semi-magically.
This required us to make an addItem method in PaneSection, so PaneSectionImpl is now abstract and has been renamed to OLAPPaneSection. All uses of the former PaneSectionImpl now support adding items.

There is one known problem with this version: if you drag something into the empty space at the bottom of a single-item pane, you get an exception. We'll look into this tomorrow.