Fixed a bug with OLAPSessions not loading with the database set.
This was happening because we put olap loading before the rest of the project when we added olap gui loading to the digester so the sql object map that was passed into the Mondrian parser was empty and each OLAPSession's database was set to null. Fixed it by passing in a new map that keeps track of OLAPSessions and the id of the database they reference and then return to set the OLAPSessions' database once the sql object map was populated.