Added primaryKey property to hierarchy edit panel. This property is required according to mondrian documentation.
This primaryKey is retrieved from the selected table.

I also discovered that it's compulsory to have a source table for a hierarchy, it's specified in either hierarchy.relation or hierarchy.primaryKeyTable. Right now we only allow a user to edit relation property.

It might be good to have a validator watching for the table selection at a later time.