Fixed a bug, "reluctant olap popup creation", popup menu are created based on selected component instead of 'clicked' component.
This required a change in the popup factor interface. 

It may be ideal if we could allow relational playpen popup factory to create popup menus for individual relational playpencomponents. But that's beyond the scope of this commit, since it's working correctly now.

OLAP playpen factory now creates popup menus from given source object, if the object is null, then an olap playpen popup is created instead.