The "bounds" property for PlayPenComponents is no longer persisted. We do not need to know the size of each component, only the location. Zoom level is independent of each client. A "location" property is now being persisted instead.
In order to draw the components correctly, it requires knowledge of the table and key names, as well as a reference to the PlayPen to calculate the preferred size of the components. The code has modified to revalidate the components when the "location" property is set or when the PlayPen is initially set in the PlayPenContentPane.