NEW - bug 2784: Have table properties dialog come before the table is actually created. http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2784
Refactored some code from TableEditPanel.applyChanges() that generates warning messages based on if the user left the table name and key name blank. This is used in the CreateTableAction class to ensure the fields in the Table Properties dialog is entered correctly before creating a new SQLTable (whereas previously, the table was created first before showing the Table Properties dialog).