Bug fix so loaded projects appear with correct type information.
The ArchitectSessionSuperConverter now looks up SQLType information in the data source collection which is different 
from the default lookup of UUIDs. More places in Architect are using the Architect specific converter than the default one.

The ArchitectSessionPersister revalidates UI components after persisting changes from the server since magic is disabled
at this time and the revalidate was not occurring.