NEW - bug 2458: Create Critic Manager http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2458
Numerous changes to the current workings of critics. At current there are two parts to the critics. The first is the
collection of classes Critic, Criticizer and Criticism. These are immutable classes that do the actual work of the critics.
The second is the CriticManager, CriticGroup and CriticSettings which hold the state of the critics and how they should
be created. Some of these classes will be merged in the near future to reduce the complexity of the critics.

The existing critics have been moved to an impl package. At current there is no way to specify settings for each critic
so there are specific implementations like OraclePhysicalNameCritic. In the future these may go away.

Added in UI components to define which critics are enabled in a project and to more easily view the critics. These settings
still need to be saved to a file or prefs although they are currently stored on the server (mainly so the server doesn't 
crash when you try to open a project).