NEW - bug 2458: Create Critic Manager http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2458
The CriticManager now handles doing the full criticizing instead of having the action store how the descendants of objects 
are traversed. The CriticManager uses a Criticizer to do the criticizing and the Criticizer has the current implementation 
of how to traverse the objects and decides which critics are given an object. In the future we may want to make different 
Criticizers for different types of object graphs to traverse.