NEW - bug 2458: Create Critic Manager http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2458
Added in a critic to check for duplicate names in support of replacing the quick fix system.
Added in a start and end method to the critics to allow critics to keep state of each run and to know when to setup and tear down its state.
Removed the getWarnings method from the DDLGenerator interface. This is the first step of cleaning up the current quick fix system but more is to come.