Added the multi-tabbed tree. Several things have been refactored:
The ArchitectFrame was changed to reference a parent context, instead of a parent session. It should now be created on its own, and have sessions added to it. It will now initialize the GUI of swing sessions, instead of the other way around.

The ArchitectSwingSession and implementation have been changed so that they no longer own a frame. They also no longer create a GUI in their constructors. Instead, initGUI is called by the frame. The constructors have changed to represent this. Note that the no argument constructor will now not create a GUI (none of them will, actually), which is opposite of previous useage. All calls have been updated.

Numerous Actions (ie. <? extends AbstractArchitectAction>) have been refactored to contain references to the frame instead of the session. Since all field accesses to session were replaced with getters, this accounts for a lot of the classes changed.

Many tests were changed to respect new constructors in the swing session.