NEW - bug 2889: Changing password doesn't work http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2889
The root cause was the rest tag was not added to the change password server request. Part of the problem was the request was nested in several anonymous classes in the UserEditorPanel. 

The request is now made in the ArchitectClientSideSession along with the rest of the server requests so they are contained in one place. In the future we may want to refactor the static server request methods to a utility class.